Implement ignoring of specific files
This is useful for progressively building out a graphql schema, but have a few files present that are placeholders.

e.g., you know you'll have a Node interface, but don't yet have a use for it, but you want to give your fellow developers a bit of guidance as to what we will have.